refactored css for artwork form components
