+++ src/App.scss
@@ -11,3 +11,7 @@
 
 @import './components/project-form/project-form';
 @import './components/project-manage/project-manage';
+@import './components/project-gallery-manage/project-gallery-manage';
+
+@import './components/artwork-form/artwork-form';
+@import './components/artwork-manage/artwork-manage';
