+++ src/components/artwork-form/_artwork-form.scss
@@ -1,0 +1,12 @@
+@import '../../stylesheets/variables';
+
+.artwork-form{
+    min-width: 40rem;
+
+    form{
+        border:1px solid $color-primary-grey
+    }
+    .padded-group{
+        padding:1rem;
+    }
+}
